Phase 11-13: Wallet/Escrow, Conditional Settlement, Reputation & Trust Engine
Phase 11 — Wallet & Escrow System:
- Wallet model (requester + agent wallets, AP Credits)
- Escrow model with audit log
- Escrow service: lock/reserve reward on task assignment
- Wallet service: seeded 5000 AP Credits for requester on startup
- WalletPage frontend with live balances

Phase 12 — Conditional Automatic Settlement:
- Settlement model + SettlementAuditLog + LedgerEntry
- Settlement service: automatic AP credit transfer on PASS
- Escrow -> releasable -> Settlement -> AP Credits transferred
- SettlementDetailsPage frontend
- Full audit trail per settlement

Phase 13 — Reputation & Trust Engine:
- ReputationEvent model (immutable RE-xxxx audit codes)
- 5-Factor deterministic formula:
  Quality 35% + Success Rate 30% + Reliability 20% + Consistency 10% + Experience 5%
- Cold-start: 80.0 baseline, Provisional tag until >=3 verified tasks
- Auto recalculation on settlement completion and verification FAIL/REVIEW
- Idempotent RE-xxxx event creation (no duplicate events)
- Reputation leaderboard, breakdown, summary, history API endpoints
- ReputationPage: live leaderboard + tier distribution + platform summary
- AgentDetailsPage: 5-factor breakdown panel + recent reputation event history
- Integration with Phase 6 Matching and Phase 7 Bid Ranking
- 63/63 integration tests passed

All phases complete: 1-13

+++ backend/app/models/bid.py
@@ -17,7 +17,7 @@ class Bid(Base):
     proposal = Column(Text, nullable=False)
     
     match_score_snapshot = Column(Float, nullable=False)
-    reputation_snapshot = Column(Integer, nullable=False)
+    reputation_snapshot = Column(Float, nullable=False)
     selection_score = Column(Float, default=0.0, nullable=False)
     
     status = Column(String(50), default="pending", nullable=False) # pending, accepted, rejected, withdrawn
